Fix TTL count for sessions with >255 test packets
Previously, a maximum of 255 counts could be represented for each
TTL value. This commit ensures that a count can be represented for
each test packet for sessions with up to the maximum allowed number
of test packets.

+++ owamp/owamp.h
@@ -1611,7 +1611,7 @@ typedef struct OWPStatsRec{
     /*
      * TTL info - histogram of received TTL values.
      */
-    uint8_t        ttl_count[256];
+    unsigned long   ttl_count[256];
 
     /*
      * Reordering buffers
